+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+TBR
+------------------
+
+* removed support for Python 3.6
+* added support for Python 3.10
+
 0.2.1 (2021-06-11)
 ------------------